Add Preferences form and view
Added a Form which allows users to display recipes based on their preferences (those include filtering out cuisines they don't like, ingredients they cannot eat and their diet preferences). For now the preferences are saved into the session variable instead of the database, and there is no verification on the inputs, only helper text to give possible values. recipes/forms.py contains methods that can be passed on in the url, that we can find useful to filter out recipes.